refactor(client) : refactor file location
according to the convention changes, file location has changed
-> all source files(*.cc) goes to it's subdir under 'srcs'. ex)  srcs/{class name}/client.cc
-> all header files(*.h) goes to the 'includes' directory. ex) includes/client.h

this location rule is decided by KenWR.
