+++ srcs/client/client.cc
@@ -1,5 +1,5 @@
 
-#include "client.h"
+#include "../../includes/client.h"
 
 namespace Just1RCe {
