+++ srcs/client/client_network_info.cc
@@ -1,7 +1,7 @@
 
 #include <sstream>
 
-#include "client.h"
+#include "../../includes/client.h"
 
 namespace Just1RCe {
